+++ traceloop-sdk/task.go
@@ -20,10 +20,10 @@ func (task *Task) LogPrompt(prompt Prompt) (LLMSpan, error) {
 	return task.workflow.sdk.LogPrompt(task.ctx, prompt, task.workflow.Attributes)
 }
 
-func (task *Task) LogAgent(attrs AgentAttributes) (LLMSpan, error) {
+func (task *Task) LogAgent(attrs AgentAttributes) LLMSpan {
 	return task.workflow.sdk.LogAgent(task.ctx, attrs, task.workflow.Attributes)
 }
 
-func (task *Task) LogToolCall(attrs ToolCallAttributes) (LLMSpan, error) {
+func (task *Task) LogToolCall(attrs ToolCallAttributes) LLMSpan {
 	return task.workflow.sdk.LogToolCall(task.ctx, attrs, task.workflow.Attributes)
 }
